# platform/datasets/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ValueGraph Datasets API",
    "description": "Stock market API with real-time and historical financial data for 27,000+ tickers over 30+ years. Financial statements, equity prices, insider trades, SEC filings, and more.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "ValueGraph"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
...
  "security": [
    {
      "X-API-KEY": []
    }
  ],
  "paths": {
...
    "/company/facts": {
      "get": {
        "tags": [
          "Company Information"
        ],
        "operationId": "getCompanyFacts",
        "summary": "Get company facts",
        "description": "Get company facts for a ticker.",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "description": "The ticker symbol.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cik",
            "in": "query",
            "description": "The CIK of the company.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company facts response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyFactsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/prices": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "operationId": "getPrices",
        "summary": "Get historical stock price data",
        "description": "Get end-of-day (EOD) historical price data for stocks.",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "description": "The stock ticker symbol (e.g. AAPL, MSFT).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "description": "The time interval for the price data.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "year"
              ]
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "The start date for the price data (format: YYYY-MM-DD).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "The end date for the price data (format: YYYY-MM-DD).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price data response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequiredError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/prices/snapshot": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "operationId": "getPriceSnapshot",
        "summary": "Price Snapshot (Real-Time)",
        "description": "Get the real-time price snapshot for a stock, including the current price, day change, and day change percent.",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "description": "The stock ticker symbol (e.g. AAPL, MSFT).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price snapshot response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceSnapshotResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequiredError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "X-API-KEY": {
        "type": "apiKey",
        "name": "X-API-KEY",
        "description": "API key for authentication.",
        "in": "header"
      }
    },
...
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A short error message."
          },
          "message": {
            "type": "string",
            "description": "A more detailed error message."
          }
        }
      },
...
      "CompanyFactsResponse": {
        "type": "object",
        "properties": {
          "company_facts": {
            "$ref": "#/components/schemas/CompanyFacts"
          }
        }
      },
...
      "PriceSnapshotResponse": {
        "type": "object",
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/PriceSnapshot"
          }
        }
      },
...
      "PricesResponse": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "description": "The ticker symbol."
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Price"
            }
          }
        }
      },
...
      "CompanyFacts": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "description": "The ticker symbol of the company."
          },
          "name": {
            "type": "string",
            "description": "The name of the company."
          },
          "cik": {
            "type": "string",
            "description": "The Central Index Key (CIK) of the company."
          },
          "industry": {
            "type": "string",
            "description": "The industry of the company."
          },
          "sector": {
            "type": "string",
            "description": "The sector of the company."
          },
          "category": {
            "type": "string",
            "description": "The category of the company."
          },
          "exchange": {
            "type": "string",
            "description": "The exchange of the company."
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the company is currently active."
          },
          "location": {
            "type": "string",
            "description": "The location of the company."
          },
          "sec_filings_url": {
            "type": "string",
            "format": "uri",
            "description": "The URL of the company's SEC filings."
          },
          "sic_code": {
            "type": "string",
            "description": "The Standard Industrial Classification (SIC) code of the company."
          },
          "sic_industry": {
            "type": "string",
            "description": "The industry of the company based on the SIC code."
          },
          "sic_sector": {
            "type": "string",
            "description": "The sector of the company based on the SIC code."
          }
        }
      },
...
      "PriceSnapshot": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "description": "The current price of the stock."
          },
          "ticker": {
            "type": "string",
            "description": "The ticker symbol."
          },
          "day_change": {
            "type": "number",
            "description": "The price change since the previous trading day's close."
          },
          "day_change_percent": {
            "type": "number",
            "description": "The percentage price change since the previous trading day's close."
          },
          "time": {
            "type": "string",
            "description": "The timestamp of the price snapshot in human-readable format in UTC."
          },
          "time_milliseconds": {
            "type": "number",
            "description": "The timestamp of the price snapshot in milliseconds since epoch."
          }
        }
      },
...
      "Price": {
        "type": "object",
        "properties": {
          "open": {
            "type": "number",
            "description": "The open price of the ticker in the given time period."
          },
          "close": {
            "type": "number",
            "description": "The close price of the ticker in the given time period."
          },
          "high": {
            "type": "number",
            "description": "The high price of the ticker in the given time period."
          },
          "low": {
            "type": "number",
            "description": "The low price of the ticker in the given time period."
          },
          "volume": {
            "type": "integer",
            "nullable": true,
            "description": "The trading volume of the ticker in the given time period."
          },
          "time": {
            "type": "string",
            "description": "The human-readable time format of the price in UTC."
          }
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Bad Request",
              "message": "Invalid request parameters"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Unauthorized",
              "message": "Invalid API key provided"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Not Found",
              "message": "Ticker XXXX not found"
            }
          }
        }
      },
      "PaymentRequiredError": {
        "description": "The request requires a paid subscription",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Payment Required",
              "message": "This endpoint requires a paid subscription. Please upgrade your plan."
            }
          }
        }
      },
...
    }
  }
}